feat: add new controller for blog user
