+++ project/apps/user/src/app/app.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common'
-import { BlogUserModule } from '@project/blog-user';
-import { AuthenticationModule } from '@project/user/authentication';
+import { BlogUserModule } from '@project/blog-user'
+import { AuthenticationModule } from '@project/authentication'
 
 @Module({
   imports: [BlogUserModule, AuthenticationModule],
