+++ project/libs/shared/helpers/src/index.ts
@@ -1,1 +1,1 @@
-export * from 'libs/shared/helpers/src/common/dto.helper'
+export { BcryptHasher } from './hasher/bcrypt.hasher'
