+++ project/libs/user/authentication/src/index.ts
@@ -1,1 +1,11 @@
 export { AuthenticationModule } from './authentication-module/authentication.module'
+export { AuthenticationService } from './authentication-module/authentication.service'
+export { AuthenticationController } from './authentication-module/authentication.controller'
+export {
+  SALT_ROUNDS,
+  DATE_FORMAT,
+  AUTH_USER_NOT_FOUND,
+  AUTH_USER_PASSWORD_WRONG,
+  AUTH_USER_EXISTS,
+  AUTH_USER_NOT_FOUN
+} from './authentication-module/authentication.constant'
