+++ project/libs/user/blog-user/src/blog-user-module/blog-user.constant.ts
@@ -1,0 +1,1 @@
+export const USER_NOT_FOUND = 'User not found'
