+++ project/libs/user/blog-user/src/blog-user-module/blog-user.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common'
-import { BaseMemoryRepository } from 'libs/shared/data-access/src/repository/base-memory.repository'
+import { BaseMemoryRepository } from '@project/data-access'
 import { BlogUserEntity } from 'libs/user/blog-user/src/blog-user-module/blog-user.entity'
 import { BlogUserFactory } from 'libs/user/blog-user/src/blog-user-module/blog-user.factory'
