+++ project/tsconfig.base.json
@@ -23,7 +23,7 @@
       "@project/data-access": ["libs/shared/data-access/src/index.ts"],
       "@project/shared/core": ["libs/shared/core/src/index.ts"],
       "@project/shared/helpers": ["libs/shared/helpers/src/index.ts"],
-      "@project/user/authentication": ["libs/user/authentication/src/index.ts"]
+      "@project/authentication": ["libs/user/authentication/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
